// src/plugins/campaign/ui/generated-admin-types.ts
export type InputMaybe<T> = T;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  JSON: { input: any; output: any; }
  Money: { input: number; output: number; }
  Upload: { input: any; output: any; }
};
// ...
export type CreateCampaignInput = {
  active: Scalars['Boolean']['input'];
  buttonText?: InputMaybe<Scalars['String']['input']>;
  defaultVariantIds: Array<InputMaybe<Scalars['String']['input']>>;
  header?: InputMaybe<Scalars['String']['input']>;
  name: Scalars['String']['input'];
  product: Scalars['ID']['input'];
  productVariants: Array<Scalars['ID']['input']>;
  subHeader?: InputMaybe<Scalars['String']['input']>;
};
// ...
export enum DeletionResult {
  /** The entity was successfully deleted */
  DELETED = 'DELETED',
  /** Deletion did not take place, reason given in message */
  NOT_DELETED = 'NOT_DELETED'
}
// ...
export type UpdateCampaignInput = {
  active?: InputMaybe<Scalars['Boolean']['input']>;
  buttonText?: InputMaybe<Scalars['String']['input']>;
  defaultVariantIds?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
  header?: InputMaybe<Scalars['String']['input']>;
  id: Scalars['ID']['input'];
  name?: InputMaybe<Scalars['String']['input']>;
  product?: InputMaybe<Scalars['ID']['input']>;
  productVariants?: InputMaybe<Array<Scalars['ID']['input']>>;
  subHeader?: InputMaybe<Scalars['String']['input']>;
};